requirements.txt
streamlit>=1.20.0,<2.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
Pillow>=9.0.0
grad-cam>=1.4.0
